requirements.txt
tensorflow-gpu==2.1.0
pyyaml
pillow
h5py==2.10.0
scikit-learn